requirements.txt
manimgl
moderngl>=5.8.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
numpy>=1.24.0
pillow>=9.0.0
